Implemented backward compatability for iCloud's file name issue.
There was an issue with wrong iCloud file name - it contained non-latin letter "A". This caused sync issues when apps where open on two devices. iOS send documentStateChanged only when app was opened/reopened.
New method fallBackFromiCloudInvalidEncodingErrorToValidURL is called when initializing iCloud access. It checks if damaged URL is present (existing users). If it is + validCloudURL is present (pre-May version of app) - it replaces content of valid doc with damaged (most recent). If only damaged url is present - app will rename damaged file to valid name.